requirements.txt
mysql-connector-python
matplotlib
Pillow
bcrypt
